requirements.txt
aiogram==2.25.1
numpy
pandas
python-dotenv
tinkoff-invest==1.0.5
requests
scikit-learn
joblib
python-telegram-bot==13.15
aiohttp
matplotlib
fastapi
uvicorn
websockets
openai